requirements.txt
# Core
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0

# Database
psycopg2-binary>=2.9,<3.0

# Cache
django-redis>=5.4,<6.0

# Logging
python-json-logger>=2.0,<3.0

# API Documentation
drf-spectacular>=0.27,<1.0

# Environment variables
python-decouple>=3.8,<4.0

# CORS
django-cors-headers>=4.3,<5.0
